--- a/lib/config/autoconfig.js
+++ b/lib/config/autoconfig.js
@@ -37,11 +37,11 @@ const MAX_CONFIG_COMBINATIONS = 17, // 16 combinations + 1 for severity only
  * @param   {number}     errorCount    The number of errors encountered when linting with the config
  */
 
- /**
-  * This callback is used to measure execution status in a progress bar
-  * @callback progressCallback
-  * @param {number} The total number of times the callback will be called.
-  */
+/**
+ * This callback is used to measure execution status in a progress bar
+ * @callback progressCallback
+ * @param {number} The total number of times the callback will be called.
+ */
 
 /**
  * Create registryItems for rules