--- a/lib/rules/comma-spacing.js
+++ b/lib/rules/comma-spacing.js
@@ -87,8 +87,8 @@ module.exports = {
 
                 },
                 message: options[dir]
-                  ? "A space is required {{dir}} ','."
-                  : "There should be no space {{dir}} ','.",
+                    ? "A space is required {{dir}} ','."
+                    : "There should be no space {{dir}} ','.",
                 data: {
                     dir
                 }