--- a/lib/rules/indent.js
+++ b/lib/rules/indent.js
@@ -468,7 +468,10 @@ module.exports = {
             context.report({
                 node: token,
                 message: createErrorMessage(neededChars, numSpaces, numTabs),
-                loc: { line: token.loc.start.line, column: token.loc.start.column },
+                loc: {
+                    start: { line: token.loc.start.line, column: 0 },
+                    end: { line: token.loc.start.line, column: token.loc.start.column }
+                },
                 fix(fixer) {
                     const range = [token.range[0] - token.loc.start.column, token.range[0]];
                     const newText = (indentType === "space" ? " " : "\t").repeat(neededChars);