--- a/lib/rules/indent.js
+++ b/lib/rules/indent.js
@@ -563,7 +563,6 @@ module.exports = {
         function addBlockIndent(node) {
 
             let blockIndentLevel;
-            const tokens = getTokensAndComments(node);
 
             if (node.parent && isOuterIIFE(node.parent)) {
                 blockIndentLevel = options.outerIIFEBody;
@@ -591,11 +590,10 @@ module.exports = {
              *     baz(); // Tokens in the block should get matched against the `if` statement, even though the opening curly is indented.
              * }
              */
+            const tokens = getTokensAndComments(node);
             const tokenToMatchAgainst = tokenInfo.isFirstTokenOfLine(tokens[0]) ? tokens[0] : sourceCode.getFirstToken(node.parent);
 
-            if (tokenToMatchAgainst !== tokens[0]) {
-                offsets.matchIndentOf(tokenToMatchAgainst, tokens[0]);
-            }
+            offsets.matchIndentOf(tokenToMatchAgainst, tokens[0]);
             offsets.setDesiredOffsets(tokens, tokens[0], blockIndentLevel);
             offsets.matchIndentOf(tokenToMatchAgainst, tokens[tokens.length - 1]);
         }