--- a/lib/rules/indent.js
+++ b/lib/rules/indent.js
@@ -8,8 +8,29 @@
 
 "use strict";
 
+//------------------------------------------------------------------------------
+// Requirements
+//------------------------------------------------------------------------------
+
 const lodash = require("lodash");
 const astUtils = require("../ast-utils");
+
+//------------------------------------------------------------------------------
+// Rule Definition
+//------------------------------------------------------------------------------
+
+/*
+ * General rule strategy:
+ * 1. An OffsetStorage instance stores a map of desired offsets, where each token has a specified offset from another
+ *    specified token or to the first column.
+ * 2. As the AST is traversed, modify the desired offsets of tokens accordingly. For example, when entering a
+ *    BlockStatement, offset all of the tokens in the BlockStatement by 1 indent level from the opening curly
+ *    brace of the BlockStatement.
+ * 3. After traversing the AST, calculate the expected indentation levels of every token according to the
+ *    OffsetStorage container.
+ * 4. For each line, compare the expected indentation of the first token to the actual indentation in the file,
+ *    and report the token if the two values are not equal.
+ */
 
 /**
  * A helper class to get token-based info related to indentation
@@ -247,10 +268,6 @@ const ELEMENT_LIST_SCHEMA = {
     ]
 };
 
-//------------------------------------------------------------------------------
-// Rule Definition
-//------------------------------------------------------------------------------
-
 module.exports = {
     meta: {
         docs: {
@@ -394,22 +411,6 @@ module.exports = {
             ObjectPattern: 1,
             flatTernaryExpressions: false
         };
-
-        /*
-         * General rule strategy:
-         * 1. Create a tree, where each node of the tree is a token or comment and each edge of the tree is an indentation offset
-         *    between tokens. (This is stored as a hashmap in `desiredOffsets`, which maps each token to an object containing
-         *    offset amount and a parent token. For example, a token in an element in an array will have
-         *    {offset: 1, from: openingBracket} to indicate that the token should be offset by 1 indentation level from the opening
-         *    bracket.) All offsets are initialized to 0 at the start.
-         * 2. As the AST is traversed, modify the offsets of tokens accordingly. For example, when entering a BlockStatement, offset
-         *    all of the tokens in the BlockStatement by 1 from the opening curly brace of the BlockStatement.
-         * 3. After traversing the AST, calculate the expected indentation levels of every token in the file according to the
-         *    `desiredOffsets` map. (Note: The indentation of some tokens with the "first" option is computed a bit differently,
-         *    using the `lockedFirstTokens` map).
-         * 4. For each token, compare the expected indentation to the actual indentation in the file, and report the token if
-         *    the two values are not equal.
-         */
 
         if (context.options.length) {
             if (context.options[0] === "tab") {