--- a/lib/rules/indent.js
+++ b/lib/rules/indent.js
@@ -722,22 +722,6 @@ module.exports = {
         }
 
         /**
-        * Checks the indentation of `IfStatement` nodes.
-        * @param {ASTNode} node An `IfStatement` node
-        * @returns {void}
-        */
-        function addIfStatementIndent(node) {
-            addBlocklessNodeIndent(node.consequent, node);
-            if (node.alternate) {
-                if (node.alternate.type === "IfStatement") {
-                    addIfStatementIndent(node.alternate, node);
-                } else {
-                    addBlocklessNodeIndent(node.alternate, node);
-                }
-            }
-        }
-
-        /**
         * Checks the indentation for nodes that are like function calls (`CallExpression` and `NewExpression`)
         * @param {ASTNode} node A CallExpression or NewExpression node
         * @returns {void}
@@ -878,7 +862,12 @@ module.exports = {
                 addFunctionParamsIndent(node, options.FunctionExpression.parameters);
             },
 
-            IfStatement: addIfStatementIndent,
+            IfStatement(node) {
+                addBlocklessNodeIndent(node.consequent, node);
+                if (node.alternate && node.alternate.type !== "IfStatement") {
+                    addBlocklessNodeIndent(node.alternate, node);
+                }
+            },
 
             ImportDeclaration(node) {
                 if (node.specifiers.some(specifier => specifier.type === "ImportSpecifier")) {