--- a/lib/rules/indent.js
+++ b/lib/rules/indent.js
@@ -672,7 +672,21 @@ module.exports = {
          */
         function addBlocklessNodeIndent(node, parent) {
             if (node.type !== "BlockStatement") {
-                offsets.setDesiredOffsets(getTokensAndComments(node), sourceCode.getFirstToken(parent), 1);
+                const firstParentToken = sourceCode.getFirstToken(parent);
+
+                offsets.setDesiredOffsets(getTokensAndComments(node), firstParentToken, 1);
+
+                /*
+                 * For blockless nodes with semicolon-first style, don't indent the semicolon.
+                 * e.g.
+                 * if (foo) bar()
+                 * ; [1, 2, 3].map(foo)
+                 */
+                const lastToken = sourceCode.getLastToken(node);
+
+                if (astUtils.isSemicolonToken(lastToken)) {
+                    offsets.matchIndentOf(firstParentToken, lastToken);
+                }
             }
         }