--- a/lib/rules/indent.js
+++ b/lib/rules/indent.js
@@ -1043,34 +1043,34 @@ module.exports = {
                         return;
                     }
 
-                    const firstTokenofLine = tokenInfo.firstTokensByLineNumber.get(lineNumber);
+                    const firstTokenOfLine = tokenInfo.firstTokensByLineNumber.get(lineNumber);
 
-                    if (firstTokenofLine.loc.start.line !== lineNumber) {
+                    if (firstTokenOfLine.loc.start.line !== lineNumber) {
 
                         // Don't check the indentation of multi-line tokens (e.g. template literals or block comments) twice.
                         return;
                     }
 
                     // If the token matches the expected expected indentation, don't report it.
-                    if (validateTokenIndent(firstTokenofLine, offsets.getDesiredIndent(firstTokenofLine))) {
+                    if (validateTokenIndent(firstTokenOfLine, offsets.getDesiredIndent(firstTokenOfLine))) {
                         return;
                     }
 
-                    if (astUtils.isCommentToken(firstTokenofLine)) {
-                        const tokenBefore = precedingTokens.get(firstTokenofLine);
+                    if (astUtils.isCommentToken(firstTokenOfLine)) {
+                        const tokenBefore = precedingTokens.get(firstTokenOfLine);
                         const tokenAfter = tokenBefore ? sourceCode.getTokenAfter(tokenBefore) : sourceCode.ast.tokens[0];
 
                         // If a comment matches the expected indentation of the token immediately before or after, don't report it.
                         if (
-                            tokenBefore && validateTokenIndent(firstTokenofLine, offsets.getDesiredIndent(tokenBefore)) ||
-                            tokenAfter && validateTokenIndent(firstTokenofLine, offsets.getDesiredIndent(tokenAfter))
+                            tokenBefore && validateTokenIndent(firstTokenOfLine, offsets.getDesiredIndent(tokenBefore)) ||
+                            tokenAfter && validateTokenIndent(firstTokenOfLine, offsets.getDesiredIndent(tokenAfter))
                         ) {
                             return;
                         }
                     }
 
                     // Otherwise, report the token/comment.
-                    report(firstTokenofLine, offsets.getDesiredIndent(firstTokenofLine));
+                    report(firstTokenOfLine, offsets.getDesiredIndent(firstTokenOfLine));
                 });
             }