--- a/lib/rules/indent.js
+++ b/lib/rules/indent.js
@@ -82,9 +82,24 @@ class OffsetStorage {
         this.tokenInfo = tokenInfo;
         this.indentType = indentType;
         this.indentSize = indentSize;
-        this.desiredOffsets = tokenInfo.getAllTokens().reduce((map, token) => map.set(token, { offset: 0, from: null }), new WeakMap());
-        this.lockedFirstTokens = new WeakMap();
-        this.desiredIndentCache = new WeakMap();
+
+        /*
+         * desiredOffsets, lockedFirstTokens, and desiredIndentCache conceptually map tokens to something else.
+         * However, they're implemented as objects with range indices as keys because this improves performance as of Node 7.
+         * This uses the assumption that no two tokens start at the same index in the program.
+         *
+         * The values of the desiredOffsets map are objects with the schema { offset: number, from: Token|null }.
+         * These objects should not be mutated or exposed outside of OffsetStorage.
+         */
+        const NO_OFFSET = { offset: 0, from: null };
+
+        this.desiredOffsets = tokenInfo.getAllTokens().reduce((desiredOffsets, token) => {
+            desiredOffsets[token.range[0]] = NO_OFFSET;
+
+            return desiredOffsets;
+        }, Object.create(null));
+        this.lockedFirstTokens = Object.create(null);
+        this.desiredIndentCache = Object.create(null);
         this.ignoredTokens = new WeakSet();
     }
 
@@ -96,7 +111,7 @@ class OffsetStorage {
     */
     matchIndentOf(baseToken, offsetToken) {
         if (baseToken !== offsetToken) {
-            this.desiredOffsets.set(offsetToken, { offset: 0, from: baseToken });
+            this.desiredOffsets[offsetToken.range[0]] = { offset: 0, from: baseToken };
         }
     }
 
@@ -117,7 +132,7 @@ class OffsetStorage {
          * element. The desired indentation of each of these tokens is computed based on the desired indentation
          * of the "first" element, rather than through the normal offset mechanism.
          */
-        this.lockedFirstTokens.set(offsetToken, baseToken);
+        this.lockedFirstTokens[offsetToken.range[0]] = baseToken;
     }
 
     /**
@@ -131,7 +146,7 @@ class OffsetStorage {
         if (offsetFrom && token.loc.start.line === offsetFrom.loc.start.line) {
             this.matchIndentOf(offsetFrom, token);
         } else {
-            this.desiredOffsets.set(token, { offset, from: offsetFrom });
+            this.desiredOffsets[token.range[0]] = { offset, from: offsetFrom };
         }
     }
 
@@ -161,31 +176,30 @@ class OffsetStorage {
     * @returns {number} The desired indent of the token
     */
     getDesiredIndent(token) {
-        if (!this.desiredIndentCache.has(token)) {
+        if (!(token.range[0] in this.desiredIndentCache)) {
 
             if (this.ignoredTokens.has(token)) {
 
                 // If the token is ignored, use the actual indent of the token as the desired indent.
                 // This ensures that no errors are reported for this token.
-                this.desiredIndentCache.set(token, this.tokenInfo.getTokenIndent(token).length / this.indentSize);
-            } else if (this.lockedFirstTokens.has(token)) {
-                const firstToken = this.lockedFirstTokens.get(token);
+                this.desiredIndentCache[token.range[0]] = this.tokenInfo.getTokenIndent(token).length / this.indentSize;
+            } else if (token.range[0] in this.lockedFirstTokens) {
+                const firstToken = this.lockedFirstTokens[token.range[0]];
 
-                this.desiredIndentCache.set(token,
+                this.desiredIndentCache[token.range[0]] =
 
                     // (indentation for the first element's line)
                     this.getDesiredIndent(this.tokenInfo.getFirstTokenOfLine(firstToken)) +
 
                         // (space between the start of the first element's line and the first element)
-                        (firstToken.loc.start.column - this.tokenInfo.getFirstTokenOfLine(firstToken).loc.start.column) / this.indentSize
-                );
+                        (firstToken.loc.start.column - this.tokenInfo.getFirstTokenOfLine(firstToken).loc.start.column) / this.indentSize;
             } else {
-                const offsetInfo = this.desiredOffsets.get(token);
+                const offsetInfo = this.desiredOffsets[token.range[0]];
 
-                this.desiredIndentCache.set(token, offsetInfo.offset + (offsetInfo.from ? this.getDesiredIndent(offsetInfo.from) : 0));
+                this.desiredIndentCache[token.range[0]] = offsetInfo.offset + (offsetInfo.from ? this.getDesiredIndent(offsetInfo.from) : 0);
             }
         }
-        return this.desiredIndentCache.get(token);
+        return this.desiredIndentCache[token.range[0]];
     }
 
     /**
@@ -205,7 +219,7 @@ class OffsetStorage {
      * @returns {Token} The token that the given token depends on, or `null` if the given token is at the top level
      */
     getFirstDependency(token) {
-        return this.desiredOffsets.get(token).from;
+        return this.desiredOffsets[token.range[0]].from;
     }
 
     /**
@@ -215,9 +229,9 @@ class OffsetStorage {
      * @returns {void}
      */
     increaseOffset(token, amount) {
-        const currentOffsetInfo = this.desiredOffsets.get(token);
+        const currentOffsetInfo = this.desiredOffsets[token.range[0]];
 
-        this.desiredOffsets.set(token, { offset: currentOffsetInfo.offset + amount, from: currentOffsetInfo.from });
+        this.desiredOffsets[token.range[0]] = { offset: currentOffsetInfo.offset + amount, from: currentOffsetInfo.from };
     }
 }