--- a/lib/rules/newline-before-return.js
+++ b/lib/rules/newline-before-return.js
@@ -50,8 +50,8 @@ module.exports = {
 
             if (node.parent.body) {
                 return Array.isArray(node.parent.body)
-                  ? node.parent.body[0] === node
-                  : node.parent.body === node;
+                    ? node.parent.body[0] === node
+                    : node.parent.body === node;
             }
 
             if (parentType === "IfStatement") {