--- a/lib/testers/rule-tester.js
+++ b/lib/testers/rule-tester.js
@@ -413,7 +413,7 @@ class RuleTester {
             const messages = result.messages;
 
             assert.equal(messages.length, 0, util.format("Should have no errors but had %d: %s",
-                        messages.length, util.inspect(messages)));
+                messages.length, util.inspect(messages)));
 
             assertASTDidntChange(result.beforeAST, result.afterAST);
         }
@@ -461,9 +461,13 @@ class RuleTester {
                 assert.equal(messages.length, item.errors, util.format("Should have %d error%s but had %d: %s",
                     item.errors, item.errors === 1 ? "" : "s", messages.length, util.inspect(messages)));
             } else {
-                assert.equal(messages.length, item.errors.length,
-                    util.format("Should have %d error%s but had %d: %s",
-                    item.errors.length, item.errors.length === 1 ? "" : "s", messages.length, util.inspect(messages)));
+                assert.equal(
+                    messages.length, item.errors.length,
+                    util.format(
+                        "Should have %d error%s but had %d: %s",
+                        item.errors.length, item.errors.length === 1 ? "" : "s", messages.length, util.inspect(messages)
+                    )
+                );
 
                 for (let i = 0, l = item.errors.length; i < l; i++) {
                     assert.ok(!("fatal" in messages[i]), `A fatal parsing error occurred: ${messages[i].message}`);