--- a/tests/bin/eslint.js
+++ b/tests/bin/eslint.js
@@ -156,9 +156,9 @@ describe("bin/eslint.js", () => {
                     assert.isTrue(fs.existsSync(CACHE_PATH), "Cache file should exist at the given location");
 
                     assert.doesNotThrow(
-                      () => JSON.parse(fs.readFileSync(CACHE_PATH, "utf8")),
-                      SyntaxError,
-                      "Cache file should contain valid JSON"
+                        () => JSON.parse(fs.readFileSync(CACHE_PATH, "utf8")),
+                        SyntaxError,
+                        "Cache file should contain valid JSON"
                     );
                 });
             });
@@ -224,9 +224,9 @@ describe("bin/eslint.js", () => {
                 return assertExitCode(child, 0).then(() => {
                     assert.isTrue(fs.existsSync(CACHE_PATH), "Cache file should exist at the given location");
                     assert.doesNotThrow(
-                      () => JSON.parse(fs.readFileSync(CACHE_PATH, "utf8")),
-                      SyntaxError,
-                      "Cache file should contain valid JSON"
+                        () => JSON.parse(fs.readFileSync(CACHE_PATH, "utf8")),
+                        SyntaxError,
+                        "Cache file should contain valid JSON"
                     );
                 });
             });