@mysticatea 2016-11-18
Amazing work, @not-an-aardvark !!

- I will review the code in this weekend.
- I think it should be a major version if `indent` rule introduces the errors of new kind of nodes by default.