@not-an-aardvark 2016-11-18
Added commits to address most of the comments.

Re. using array join in the multiline test cases, I'm thinking maybe we should try to improve the readability of these tests in general. How would you feel about doing something like this?

```js
ruleTester.run("indent", rule, {
    valid: [
        code: `
if (foo) {
  bar();
} else {
  qux();
}
`,
        options: [2]
    ]
});
```

That would allow the tests to clearly represent indentation without boilerplate from array joins or string concatenation, but it does make the file look less nice overall.

We could also use tagged template literals to avoid this effect by removing a constant number of indentation spaces from the resulting string. However, that might make it harder to see what's going on without looking at the template function.

```js
function unIndent(strings) {
  // remove a constant amount of indentation, and return the resulting string.
}

ruleTester.run("indent", rule, {
    valid: [
        code: unIndent`
            if (foo) {
              bar();
            } else {
              qux();
            }
        `, // this might be a bit misleading about how much indentation the string actually has
        options: [4]
    ]
});
```