@platinumazure 2016-11-18
Re: readability of multiline tests: I'm certainly not strongly attached to array join. I'm open to suggestions (such as a normal template string or a tagged template string).

I just find the `\n` at the end of the test case lines distracting, as it makes it harder to spot a trailing punctuator token (e.g., opening curly brace). As far as I'm concerned, nearly anything else is an improvement over that.