@lukeapage 2016-11-19
Does this ignore jsx and continue to leave that to the react plugin jsx indent rule?