@not-an-aardvark 2016-11-19
@lukeapage I'm not sure I understand the question. Could you clarify what you're expecting the new implementation to do?