@lukeapage 2016-11-19
So,  either jsx e.g. The lines with html tags are checked now and therefore the react plugin rule becomes obsolete, or those lines are also excluded and people should continue to use the rule in the react plugin to verify jsx indentation. 
I just wanted to make sure whatever approach was part of the plan.