@not-an-aardvark 2016-11-19
Thanks for bringing this up. I am not very familiar with JSX, and this implementation does not explicitly account for it. However, given that every token in a file is now checked, I think it's likely that the new implementation might report errors that the old implementation did not.

Maybe we should add special logic to account for unknown AST nodes.