@lukeapage 2016-11-19
heres the react plugin rule:
https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/jsx-indent.md
because you can put standard js inside jsx you can end up with quite complex code to check e.g.
```
return a && (
    <a>    
        {
             b ?
                "string" :
               <more-jsx/>
        }
    </a>);
```
etc.
you can see the number of different scenarios in the test suite for the react plugin
https://github.com/yannickcr/eslint-plugin-react/blob/master/tests/lib/rules/jsx-indent.js

so maybe excluding everything inside an unknown node might work?