@not-an-aardvark 2016-11-19
Hmm, on second thought I think excluding unknown AST nodes would not be a good idea, because whenever new syntax became supported the rule would end up unexpectedly ignoring it, which is the same problem we have now.