@vitorbal 2016-11-19
Maybe we can just ignore indentation inside `JSXElement` nodes?