@not-an-aardvark 2016-11-22
@platinumazure No problem! Please keep pointing out misleading comments if you find them; they'll make it difficult to understand the code in the future.