@mysticatea 2016-11-23
I'm sorry for late.

Again, I appreciate indent rule's rewrite. Thank you!

I found some issues.

## Stack overflow

I got a stack overflow error if I lint [acorn/src/expression.js](https://github.com/ternjs/acorn/blob/3814997411c4ca829fd91bc33d4271ee08bdf2ba/src/expression.js).
The current indent rule can finish linting the file.

```
$ cd acorn
$ npm install ../eslint
$ ./node_modules/.bin/eslint src/expression.js --rule "indent:[error,2]" --parser-options "ecmaVersion:6,sourceType:module"
```

## Performance

I saw that the new implementation was x30 slower in some my private repository.
I'm sorry, I cannot show the repositories...
I guess this problem will be solved if the stack overflow problem is solved.

    eslint src --no-eslintrc --env es6 --rule "indent:error"

- Before: 4 seconds
- After: 119 seconds

## Exports

I got unexpected errors on export statements of ES Modules.

```js
/*eslint indent: error */
const foo = 1, bar = 2;

export {
    foo,  //※※※ error  Expected indentation of 0 spaces but found 4  indent
    bar   //※※※ error  Expected indentation of 0 spaces but found 4  indent
}
```

## Template curlies

I felt that handling template curlies of this implementation was unexpected a bit.
It seems:

- The first line in template curlies is expected 1 indent from the template's indent.
- The closing curly of template curlies is expected the same indent as the template's indent.

This behavior works perfectly for short templates, but not for long templates.
An example is [eaw/scripts/generate.js](https://github.com/mysticatea/eaw/blob/881acec64ad4a1703aa154dba2fe5a1393dac6eb/scripts/generate.js#L133-L174); I expected no error in this code, but got 2 errors.

```
  161:13  error  Expected indentation of 8 spaces but found 12  indent
  168:9   error  Expected indentation of 4 spaces but found 8   indent

✖ 2 problems (2 errors, 0 warnings)
```

People might use long template literals to generate HTML or something like.
In that case, indentation which is based on the indent of the opening curly is expected because the start of the template literal is too far, I think.

This behavior is complex and there are vary preferences.
Maybe it should not warn the first line in template curlies and the closing curly.

## Semicolons at the first of lines

```js
const doSomething = () => {}
;[1,2,3].forEach(doSomething)  //※※※ error  Expected indentation of 4 spaces but found 0  indent
```

I think this is an unexpected error for the semicolon-less style.

## Nested ternary.

I like the following style:

```js
let foo = (
    a === 1 ? "one" :
    a === 2 ? "two" :
    a === 3 ? "three" :
    a === 4 ? "four" :
    /* otherwise */ "more"
)
```

The current indent rule allows this code, but this implementation disallows.
I want to allow the code above. I'm OK if it's behind of an option.