@mysticatea 2016-11-24
I rethought and edited the section about "template curlies".