@not-an-aardvark 2016-11-24
@mysticatea Thanks for the review! I'll try to fix the issues that you mentioned.

@platinumazure I've been working on making the test file easier to read. How does [this](https://github.com/not-an-aardvark/eslint/blob/02f0c565ce52107bf7468c40bd492fa3856b309a/tests/lib/rules/indent.js) look? (I pushed it onto a separate branch for now to avoid making the diff for this PR too large, but let me know if I should merge it into this PR.)