@not-an-aardvark 2016-11-30
@mysticatea How should the rule handle this?

```js
var foo = () => {},
    bar = 1,
    baz = 2
    ; // <-- The rule currently enforces a semicolon here

var foo = () => {},
    bar = 1,
    baz = 2
    ; [1, 2, 3].map(foo) // <-- This looks strange

var foo = () => {},
    bar = 1,
    baz = 2
; [1, 2, 3].map(foo) // <-- This looks ok, but it contradicts the first option.
```

I think it would be best to make the first option incorrect, so the semicolon always has no offset.

```js
var foo = () => {},
    bar = 1,
    baz = 2
;
```