@mysticatea 2016-12-04
- ✔ Stack overflow
- ✔ Exports
- ✔ Template curlies

## Performance

In my private repo, new indent rule still seems to be x135 slower than the current indent rule.
It seems to get so slow if there are large object literals or array literals, such as [test-data.js](https://gist.github.com/mysticatea/2fca6bc8c2194f75e4808a884a1e351a).

- Before: 803.928 ms
- After: 115286.021 ms

In eslint repo, new indent rule seems to be x6 slower than the current indent rule.

- Before: 697.510 ms
- After: 4314.090 ms

<details>

```
C:\Users\t-nagashima.AD\Documents\GitHub\eslint [indent-rewrite ≡]> eslint lib --env es6 --rule "indent:[error,4,{SwitchCase:1}]" --no-eslintrc
Rule                 | Time (ms) | Relative
:--------------------|----------:|--------:
indent               |  4436.618 |   100.0%
no-underscore-dangle |     0.407 |     0.0%
C:\Users\t-nagashima.AD\Documents\GitHub\eslint [indent-rewrite ≡]> eslint lib --env es6 --rule "indent:[error,4,{SwitchCase:1}]" --no-eslintrc
Rule                 | Time (ms) | Relative
:--------------------|----------:|--------:
indent               |  4277.314 |   100.0%
no-underscore-dangle |     0.399 |     0.0%
C:\Users\t-nagashima.AD\Documents\GitHub\eslint [indent-rewrite ≡]> eslint lib --env es6 --rule "indent:[error,4,{SwitchCase:1}]" --no-eslintrc
Rule                 | Time (ms) | Relative
:--------------------|----------:|--------:
indent               |  4223.935 |   100.0%
no-underscore-dangle |     0.379 |     0.0%
C:\Users\t-nagashima.AD\Documents\GitHub\eslint [indent-rewrite ≡]> eslint lib --env es6 --rule "indent:[error,4,{SwitchCase:1}]" --no-eslintrc
Rule                 | Time (ms) | Relative
:--------------------|----------:|--------:
indent               |  4417.716 |   100.0%
no-underscore-dangle |     0.396 |     0.0%
C:\Users\t-nagashima.AD\Documents\GitHub\eslint [indent-rewrite ≡]> eslint lib --env es6 --rule "indent:[error,4,{SwitchCase:1}]" --no-eslintrc
Rule                 | Time (ms) | Relative
:--------------------|----------:|--------:
indent               |  4314.090 |   100.0%
no-underscore-dangle |     0.407 |     0.0%
C:\Users\t-nagashima.AD\Documents\GitHub\eslint [indent-rewrite ≡]> git co master
Switched to branch 'master'
Your branch is up-to-date with 'origin/master'.
C:\Users\t-nagashima.AD\Documents\GitHub\eslint [master ≡]> eslint lib --env es6 --rule "indent:[error,4,{SwitchCase:1}]" --no-eslintrc
Rule                 | Time (ms) | Relative
:--------------------|----------:|--------:
indent               |   680.669 |    99.9%
no-underscore-dangle |     0.395 |     0.1%
C:\Users\t-nagashima.AD\Documents\GitHub\eslint [master ≡]> eslint lib --env es6 --rule "indent:[error,4,{SwitchCase:1}]" --no-eslintrc
Rule                 | Time (ms) | Relative
:--------------------|----------:|--------:
indent               |   697.510 |    99.9%
no-underscore-dangle |     0.400 |     0.1%
C:\Users\t-nagashima.AD\Documents\GitHub\eslint [master ≡]> eslint lib --env es6 --rule "indent:[error,4,{SwitchCase:1}]" --no-eslintrc
Rule                 | Time (ms) | Relative
:--------------------|----------:|--------:
indent               |   699.458 |    99.9%
no-underscore-dangle |     0.413 |     0.1%
C:\Users\t-nagashima.AD\Documents\GitHub\eslint [master ≡]> eslint lib --env es6 --rule "indent:[error,4,{SwitchCase:1}]" --no-eslintrc
Rule                 | Time (ms) | Relative
:--------------------|----------:|--------:
indent               |   706.844 |    99.9%
no-underscore-dangle |     0.422 |     0.1%
C:\Users\t-nagashima.AD\Documents\GitHub\eslint [master ≡]> eslint lib --env es6 --rule "indent:[error,4,{SwitchCase:1}]" --no-eslintrc
Rule                 | Time (ms) | Relative
:--------------------|----------:|--------:
indent               |   677.977 |    99.9%
no-underscore-dangle |     0.405 |     0.1%
```

I'm not sure why there are `no-underscore-dangle`.

</details>

.
I felt we need to improve this performance.


## Semicolons at the first of lines

I think another option is:

- 1 indent if there are statements preceded by the semicolon.
- 0 indent if the semicolon is only token on the line.

```js
// ✔ GOOD
var foo = () => {}
  , bar = 1
  , baz = 2
  ;
var foo = () => {}
  , bar = 1
  , baz = 2
;[bar, baz].forEach(foo)

// ✘ BAD
var foo = () => {}
  , bar = 1
  , baz = 2
;
var foo = () => {}
  , bar = 1
  , baz = 2
  ;[bar, baz].forEach(foo)
```